requirements.txt
Flask>=2.3.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.23.0
diffusers>=0.21.0
transformers>=4.30.0
huggingface-hub>=0.16.0
tqdm>=4.65.0
accelerate>=0.20.0
safetensors>=0.3.0
